+++ error-messages/user.messages.js
@@ -12,9 +12,11 @@ const AUTHENTICATION_TOKEN_NOT_VALID = 'AUTHENTICATION_TOKEN_NOT_VALID';
 const RESET_PASSWORD_TOKEN_NOT_VALID = 'RESET_PASSWORD_TOKEN_NOT_VALID';
 const USER_EMAIL_ALREADY_CONFIRMED = 'USER_EMAIL_ALREADY_CONFIRMED';
 const INVALID_ADMIN_INVITATIONAL_TOKEN = 'INVALID_ADMIN_INVITATIONAL_TOKEN';
+const ID_NOT_PROVIDED = 'ID_NOT_PROVIDED';
 
 module.exports = {
   USER_ALREADY_EXIST,
+  ID_NOT_PROVIDED,
   USER_NOT_FOUND,
   INPUT_NOT_VALID,
   WRONG_CREDENTIALS,
