+++ modules/user/user.graphql.js
@@ -23,7 +23,6 @@ type User{
 
 const userInput = `
 input UserInput {
-    _id:ID
     firstName: String
     lastName: String
     password: String
